Implemented basic funtionality.
Not yet tested.  This commit should have the basic functionality
required to connect, initialize video streaming, and stop streaming.

Rendering implementation has started, however it is not complete.
Considering changing the Isonchronous Packet Implementation to include
packet parsing and frame building in the same thread that listens and
receives packets.  This will reduce the number of copies necessary, and
should have no affect on the speed in which USB Requests are received.