Stub Renderer class.  Submit ByteBuffer packet implementation
ByteBuffer implementation seems to drop packets.  They aren't coming in
the order the driver would expect, causing nearly all frames to drop.  I
don't know if its due to the speed of the implementation or if its due
to an unknown result of the WeakReference that JNA keeps to ByteBuffers.

Regardless, next I will attempt the base implentation of just reading
the UsbIso buffer into a byte array. I will immediately send it to a
handler for processing.