Implemented local frame pool.
While I am committing this for archival purposes, I believe a better way
of handling this is to create direct bytebuffers in JNI and send those
references back to java.  This should reduce processing time as I only
need to callback to Java once, rather than twice (once to get the
buffer, then again to return it after contents are copied to it).  It
will also signifcantly reduce the memory footprint.