Moved Java renderer to Example App.
Tested a Bitmap implementation for the renderer, it was far slower than
just using a byte array, as I should have expected.

The Java based renderer was moved because it has no place in the
library.  Since I'm already working in JNI for the library all included
renderers should be implemented in native code, its faster and makes
more sense.