Add support for older devices.
The oldest device I can test on is API 15, so that is the minimum I have
set.  Technically I believe I could support api levels lower than that,
as long as they have support for otg/usb host.

Sample2 is required for older devices (GLES Renderer).

There is something a bit wrong with the GL Renderer.  Occasionally for
the first 5-10 seconds there is a slight color swirling affect on
screen.  It corrects itself, but I am unsure as to what the cause is.  I
didn't notice it during renderscript testing, but perhaps its because I
wasn't testing on a larger (tablet) screen.