Tweaked GL Renderer(s).
Changed the TEXTURE_MIN and MAG filters to GL_NEAREST for the mask. That
may be what was causing the "swirling" artifact, but it needs more
testing.  If the linear averages of the mask position were close and
changing, it could have been altering per frame which y-value was being
used.