Some refactoring in prerparation of implementing async bulk transfers.
I also finally found a real fix for the swirling problem.  Setting a
fixed size for the OpenGL surface forces hardware scaling.  This sets
the OpenGL viewport to the exact size of my frame, which allows me to
get rid of Texture mask I used to determine the y-Value, as I can simple
use gl_FragCoord.

This commit only implements this in the Java OGLRenderer used in sample
2.  I have started to rework the Native Renderer, but I am not entirely
sure I want to keep it around.  There doesn't seem to be a significant
speed difference.