Make sure device isn't streaming before setting listener.
I decided to keep the FrameListener functionality the way it currently
is, but added a check to make sure that the device isn't streaming (thus
no callbacks are executed) before allowing the command to proceed.  I
can do this because the commands to the driver themselves are thread
safe.  I know with certaintly that a start or stop cannot be issued to
the device simultaneously with setting the listener.

Currently onError is called if the device is streaming, however in the
future I can stop/restart streaming rather than issue an error.