Updated to NDK r16b.  Fixed some bugs.
Tested Sample 2 on real hardware (a Nexus 7).  There are rough edges
that need addressing along with more thorough testing, but in general
video capture works and is rendered in OpenGL.

Bugs Fixed:
1)  Constant literal with an f appendix wouldn't compile for GLSL.  This
apparently only an issue for older hardware

2) Fixed a potentential deadlock where its possible for the frame pool
to spin forever whent the device is closed.

Also changed the native frame lock from an atomic_counter back to an
atomic flag.  The counter was necessary when it was possible for
multiple threads to consume a frame, but an atomic_flag is a bit more
intuitive for a lock and is also a bit faster as it doesn't require
fencing.  Acquire/Release memory order also makes it more reliable.